Fix invalid markup generated for dropdown in Bootstrap 4
When run in Bootstrap 4, the plugin creates <a> elements for the dropdown.
However, this is invalid markup, as the href attribute is required for the <a>
element. The result of this omission is that the text color is incorrect in
many themes and the cursor is not a pointer.

This pull request fixes the oversight.